Restructure processors
