+++ internal/processing/istio/reconciliation.go
@@ -5,6 +5,7 @@ import (
 
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
 	"github.com/kyma-incubator/api-gateway/internal/processing"
+	"github.com/kyma-incubator/api-gateway/internal/processing/processors"
 	"github.com/kyma-incubator/api-gateway/internal/validation"
 	networkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -17,17 +18,17 @@ type Reconciliation struct {
 
 func NewIstioReconciliation(config processing.ReconciliationConfig) Reconciliation {
 	vsProcessor := NewVirtualServiceProcessor(config)
-	apProcessor := NewAuthorizationPolicyProcessor(config)
-	raProcessor := NewRequestAuthenticationProcessor(config)
+	apProcessor := processors.NewAuthorizationPolicyProcessor(config)
+	raProcessor := processors.NewRequestAuthenticationProcessor(config)
+	acProcessor := processors.NewAccessRuleProcessor(config)
 
 	return Reconciliation{
-		processors: []processing.ReconciliationProcessor{vsProcessor, raProcessor, apProcessor},
+		processors: []processing.ReconciliationProcessor{vsProcessor, raProcessor, apProcessor, acProcessor},
 		config:     config,
 	}
 }
 
 func (r Reconciliation) Validate(ctx context.Context, client client.Client, apiRule *gatewayv1beta1.APIRule) ([]validation.Failure, error) {
-
 	var vsList networkingv1beta1.VirtualServiceList
 	if err := client.List(ctx, &vsList); err != nil {
 		return make([]validation.Failure, 0), err
