+++ internal/processing/processors/authorization_policy_processor_test.go
@@ -1,4 +1,4 @@
-package istio_test
+package processors_test
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
 	. "github.com/kyma-incubator/api-gateway/internal/processing/internal/test"
-	"github.com/kyma-incubator/api-gateway/internal/processing/istio"
+	"github.com/kyma-incubator/api-gateway/internal/processing/processors"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	securityv1beta1 "istio.io/client-go/pkg/apis/security/v1beta1"
@@ -39,7 +39,7 @@ var _ = Describe("Authorization Policy Processor", func() {
 		ruleJwt2 := GetRuleWithServiceFor(ImgApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt}, service)
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt, ruleJwt2})
 		client := GetEmptyFakeClient()
-		processor := istio.NewAuthorizationPolicyProcessor(GetTestConfig())
+		processor := processors.NewAuthorizationPolicyProcessor(GetTestConfig())
 
 		// when
 		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
@@ -102,7 +102,7 @@ var _ = Describe("Authorization Policy Processor", func() {
 		client := GetEmptyFakeClient()
 		ruleJwt := GetRuleFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt})
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
-		processor := istio.NewAuthorizationPolicyProcessor(GetTestConfig())
+		processor := processors.NewAuthorizationPolicyProcessor(GetTestConfig())
 
 		// when
 		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
@@ -127,8 +127,7 @@ var _ = Describe("Authorization Policy Processor", func() {
 		client := GetEmptyFakeClient()
 		ruleJwt := GetRuleWithServiceFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt}, service)
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
-
-		processor := istio.NewAuthorizationPolicyProcessor(GetTestConfig())
+		processor := processors.NewAuthorizationPolicyProcessor(GetTestConfig())
 
 		// when
 		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
@@ -161,7 +160,7 @@ var _ = Describe("Authorization Policy Processor", func() {
 		}
 		ruleJwt := GetRuleWithServiceFor(HeadersApiPath, ApiMethods, []*gatewayv1beta1.Mutator{}, []*gatewayv1beta1.Authenticator{jwt}, service)
 		apiRule := GetAPIRuleFor([]gatewayv1beta1.Rule{ruleJwt})
-		processor := istio.NewAuthorizationPolicyProcessor(GetTestConfig())
+		processor := processors.NewAuthorizationPolicyProcessor(GetTestConfig())
 
 		// when
 		result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
