+++ internal/processing/processors/request_authentication_processor.go
@@ -1,4 +1,4 @@
-package istio
+package processors
 
 import (
 	"context"
