+++ internal/processing/processors/virtual_service_processor.go
@@ -1,8 +1,10 @@
-package processing
+package processors
 
 import (
 	"context"
+
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
+	"github.com/kyma-incubator/api-gateway/internal/processing"
 	networkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -17,16 +19,16 @@ type VirtualServiceCreator interface {
 	Create(api *gatewayv1beta1.APIRule) *networkingv1beta1.VirtualService
 }
 
-func (r VirtualServiceProcessor) EvaluateReconciliation(ctx context.Context, client ctrlclient.Client, apiRule *gatewayv1beta1.APIRule) ([]*ObjectChange, error) {
+func (r VirtualServiceProcessor) EvaluateReconciliation(ctx context.Context, client ctrlclient.Client, apiRule *gatewayv1beta1.APIRule) ([]*processing.ObjectChange, error) {
 	desired := r.getDesiredState(apiRule)
 	actual, err := r.getActualState(ctx, client, apiRule)
 	if err != nil {
-		return make([]*ObjectChange, 0), err
+		return make([]*processing.ObjectChange, 0), err
 	}
 
 	changes := r.getObjectChanges(desired, actual)
 
-	return []*ObjectChange{changes}, nil
+	return []*processing.ObjectChange{changes}, nil
 }
 
 func (r VirtualServiceProcessor) getDesiredState(api *gatewayv1beta1.APIRule) *networkingv1beta1.VirtualService {
@@ -34,7 +36,7 @@ func (r VirtualServiceProcessor) getDesiredState(api *gatewayv1beta1.APIRule) *n
 }
 
 func (r VirtualServiceProcessor) getActualState(ctx context.Context, client ctrlclient.Client, api *gatewayv1beta1.APIRule) (*networkingv1beta1.VirtualService, error) {
-	labels := GetOwnerLabels(api)
+	labels := processing.GetOwnerLabels(api)
 
 	var vsList networkingv1beta1.VirtualServiceList
 	if err := client.List(ctx, &vsList, ctrlclient.MatchingLabels(labels)); err != nil {
@@ -48,11 +50,11 @@ func (r VirtualServiceProcessor) getActualState(ctx context.Context, client ctrl
 	}
 }
 
-func (r VirtualServiceProcessor) getObjectChanges(desiredVs *networkingv1beta1.VirtualService, actualVs *networkingv1beta1.VirtualService) *ObjectChange {
+func (r VirtualServiceProcessor) getObjectChanges(desiredVs *networkingv1beta1.VirtualService, actualVs *networkingv1beta1.VirtualService) *processing.ObjectChange {
 	if actualVs != nil {
 		actualVs.Spec = *desiredVs.Spec.DeepCopy()
-		return NewObjectUpdateAction(actualVs)
+		return processing.NewObjectUpdateAction(actualVs)
 	} else {
-		return NewObjectCreateAction(desiredVs)
+		return processing.NewObjectCreateAction(desiredVs)
 	}
 }
